[PlaybackSerialiser] Added PlaybackExtensions
Playback extensions are an easy way to add more functionality to the
Playback file and controller, without creating a new flavor or extra events.

- [PlaybackController] Changed index from int to long
- [PlaybackController] Added CommentHandler
- [PlaybackController] Added EventRecordTick and EventPlaybackTick

+++ src/main/java/com/minecrafttas/tasmod/gui/InfoHud.java
@@ -410,7 +410,7 @@ public class InfoHud extends GuiScreen implements EventClientTick, EventDrawHotb
 			if (configuration.getProperty(title + "_x", "err").equals("err")) setDefaults(title, y, true);
 			lists.add(new InfoLabel(title, Integer.parseInt(configuration.getProperty(title + "_x")), Integer.parseInt(configuration.getProperty(title + "_y")), Boolean.parseBoolean(configuration.getProperty(title + "_visible")), Boolean.parseBoolean(configuration.getProperty(title + "_rect")), () -> {
 				if (Minecraft.getMinecraft().currentScreen == this) return "PlaybackIndex";
-				return Integer.toString(TASmodClient.controller.index());
+				return Long.toString(TASmodClient.controller.index());
 			}));
 			
 			y = height - 14;
